🚚 rename: ReasonSelect를 컴포넌트 index로 변경 #42
